update tutorial 9 and example cmakelist
