+++ examples/simple_plugin/CMakeLists.txt
@@ -6,8 +6,7 @@ find_package(ignition-plugin${IGN_PLUGIN_VER} 1.1 REQUIRED COMPONENTS all)
 set(IGN_PHYSICS_VER 2)
 find_package(ignition-physics${IGN_PHYSICS_VER} REQUIRED)
 
-add_library(SimpleTPEPlugin SHARED plugin.cc)
-target_link_libraries(SimpleTPEPlugin
+add_library(SimplePlugin SHARED plugin.cc)
+target_link_libraries(SimplePlugin
   PRIVATE
-    ignition-plugin${IGN_PLUGIN_VER}::loader
     ignition-physics${IGN_PHYSICS_VER}::ignition-physics${IGN_PHYSICS_VER})
